FIX: Removed an id as Query Param

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/LocationService.java
@@ -9,5 +9,5 @@ public interface LocationService {
 
     Location findById(Long locationId);
 
-    List<Location> getLocationByPublicationId(Integer publicationId);
+    List<Location> getLocationByPublicationId(Long publicationId);
 }
